// src/Infogroup.IDMS.Web.Host/src/app/main/campaigns/shared/campaign-action.enum.ts
export enum CampaignAction {
    SubmitOutput = 'SUBMIT_OUTPUT',
    SubmitOrder = 'SUBMIT_ORDER',
    Ship = 'SHIP',
    Reship = 'RESHIP',
    Delete = 'DELETE',
    CancelCampaign = 'CANCEL_CAMPAIGN',
    SaveSelection = 'SAVE_SELECTION',
    ScheduleCampaign = 'SCHEDULE_CAMPAIGN',
    ScheduleOutput = 'SCHEDULE_OUTPUT',
    OutputLayout = 'OUTPUT_LAYOUT',
    ResetShip = 'RESET_SHIP'

}