// src/Infogroup.IDMS.Web.Host/src/app/main/campaigns/shared/campaign-status.enum.ts
export enum CampaignStatus {
    CampaignCreated = 10,
    CampaignSubmitted = 20,
    CampaignRunning = 30,
    CampaignCompleted = 40,
    CampaignFailed = 50,
    ReadytoOutput = 60,
    OutputSubmitted = 70,
    OutputRunning = 80,
    OutputCompleted = 90,
    OutputFailed = 100,
    ApprovedforShipping = 110,
    WaitingtoShip = 120,
    Shipped = 130,
    ShippingFailed  = 140,
    Cancelled = 150
}